Fix auipcc RV32 representability issue
This extends the representable bound to a MW-bit value and
gives precisely 1/4 of the representable region below the base.
The result is to ensure ojects >2k always have 2k below and
above the bounds, ensuring auipcc is representable if a 12-bit
signed immediate is to then be used to bring the object back
into bounds.